tracing: Allow Android to use blktrace without debugfs
Android 11's IORap prefetching service requires blktrace to build an
index of what data each app accesses while it's launching so that it
can prefetch said data as soon as the launching process starts, before
the app actually needs the data. This results in significantly reduced
cold app launch times when the required data is not present in the page
cache, e.g. 981 -> 936 ms (-5%) even for simple apps such as Magisk
Manager. According to Google, larger real-world apps can launch up to
26% faster depending on the device and app in question [1]. The
profiling overhead introduced by tracing has been measured to be
negligible.

Android only needs basic blktrace functionality through tracefs, and the
core of the tracing subsystem only uses debugfs to mount tracefs, so we
can unlock tracing and blktrace for it to use with minimal dependencies
on debugfs.

[1] https://medium.com/androiddevelopers/improving-app-startup-with-i-o-prefetching-62fbdb9c9020

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>